Feature/mejorar deteccion (#43)
* chore(pre-commit): reestructurar configuración de hooks y actualizar entradas para Ruff, Black y MyPy

* feat(detección): agregar registro de eventos y manejo de errores en el procesamiento de videos y API

- Se implementó un sistema de logging en las clases DetectionApp, DetectionAPI, DetectorService, VideoProcessor y ZoneList.
- Se mejoró el manejo de excepciones para registrar errores críticos y advertencias.
- Se agregó información de estado en el procesamiento de videos y en la activación de multas.

* WIP: Intento de solución de errores:
- ✅ Mejorar los logs ya que no se entienden los errores
- Falla la detección al utilizar la cámara, cuando detecta un vehículo
- El factor de escala de "zones" no se calcula bien o hay algún problema. La idea es que las zonas están delimitadas en videos 1080x1920 y se deberían adaptar en base a la resolución del video que se muestra.
- El cerrado del sistema no funciona correctamente. Ante una falla, el sistema permanece levantado.
- Al hacer ctrl + c, el sistema no se cierra correctamente.
- Al intentar cerrar una ventana de video, se vuelve a abrir otra ventana pero de dimensión igual al video que está reproduciendo.
- Video_processor utiliza una ruta qumada en el código en vez de una de configuración: result_folder = os.path.join(os.getcwd(), "detection_results")

* feat: Implementar clases de fuentes de stream y gestor de ventanas

- Añadir `StreamSource`, `VideoStreamSource` y `CameraStreamSource` para manejar entradas de video y cámara.
- Implementar `WindowManager` para gestionar la visualización de ventanas con OpenCV.
- Crear pruebas unitarias para verificar la arquitectura de clases y la funcionalidad de los nuevos componentes.
- Añadir pruebas de integración para asegurar que la refactorización no rompe la funcionalidad existente.
- Incluir pruebas funcionales para validar el procesamiento de videos reales.

* fix(detection): Sistema de detección mejorado con superposición profesional y nuevos endpoints de API

- Se añadió un nuevo endpoint de API para recuperar tiempos de espera en zonas específicas.
- Se introdujo la clase InfoOverlay para la visualización profesional de información en vídeos de detección.
- Se actualizó StreamCoordinator para gestionar las devoluciones de llamadas de FPS y la integración de superposiciones.
- Se mejoró la organización del código eliminando comentarios innecesarios y reestructurando archivos.
- Se eliminó la configuración de zona 4K obsoleta de YAML.
- Se añadió la colección Postman para pruebas y documentación de API.
- Se garantizó el cumplimiento de las mejores prácticas y principios de codificación (DRY, SOLID, KISS).

* fix(detección): actualizar puntos de zonas y normalizar tiempos de detección por FPS

* feat(config): actualizar configuración de servicios y detección

- Se agregó un nuevo puerto de detección en la configuración.
- Se modificaron las rutas de origen y destino para el procesamiento de videos.
- Se ajustaron las configuraciones de generación de reportes y simulación.
- Se implementaron validaciones en los modelos de datos del reporte.
- Se refactorizó el servicio de reportes para utilizar el nuevo modelo de datos.
- Se mejoró la lógica de generación y almacenamiento de reportes, incluyendo validaciones y manejo de errores.

* WIP: Actualización a python 3.12 y de tensorflow

* WIP: Agregar supervision y ultralytics en pyproject.toml

* fix(sincronización): mejorar manejo de errores y respuesta en el estado de sincronización

* WIP: Arreglar tamaño de ventana en deteccion

* fix(api): eliminar impresión de reporte generado en SumoAPI

* WIP: Arreglar "reporte"

- Renombrar columnas
- Fix alerta.log

* refactor: Agregar carriles a la simulacion y agregar steps customs en decision

* Agrupar backup_dependencies

* feat(types): agregar tipo de resolución y actualizar referencias en el sistema

- Se agregó un nuevo tipo `Resolution` para representar resoluciones de video/pantalla.
- Se actualizaron las funciones y clases para utilizar el nuevo tipo `Resolution` en lugar de tuplas.
- Se mejoró la legibilidad del código al usar parámetros nombrados en las consultas SQL.

+++ src/traffic_system/decision/DQN/evaluation_metrics.py
@@ -177,13 +177,13 @@ class DQNEvaluator:
 
         # Calcular estadísticas
         metrics = {
-            "mean_reward": np.mean(evaluation_rewards),
-            "std_reward": np.std(evaluation_rewards),
-            "mean_waiting_time": np.mean(evaluation_waiting_times),
-            "std_waiting_time": np.std(evaluation_waiting_times),
-            "mean_throughput": np.mean(evaluation_throughput),
-            "std_throughput": np.std(evaluation_throughput),
-            "mean_episode_length": np.mean(evaluation_episode_lengths),
+            "mean_reward": float(np.mean(evaluation_rewards)),
+            "std_reward": float(np.std(evaluation_rewards)),
+            "mean_waiting_time": float(np.mean(evaluation_waiting_times)),
+            "std_waiting_time": float(np.std(evaluation_waiting_times)),
+            "mean_throughput": float(np.mean(evaluation_throughput)),
+            "std_throughput": float(np.std(evaluation_throughput)),
+            "mean_episode_length": float(np.mean(evaluation_episode_lengths)),
             "total_episodes": num_episodes,
         }
 
@@ -194,12 +194,14 @@ class DQNEvaluator:
             else 0
         )
         self.evaluation_metrics["evaluation_episodes"].append(current_episode)
-        self.evaluation_metrics["evaluation_rewards"].append(metrics["mean_reward"])
+        self.evaluation_metrics["evaluation_rewards"].append(
+            float(metrics["mean_reward"])
+        )
         self.evaluation_metrics["evaluation_waiting_times"].append(
-            metrics["mean_waiting_time"]
+            float(metrics["mean_waiting_time"])
         )
         self.evaluation_metrics["evaluation_throughput"].append(
-            metrics["mean_throughput"]
+            float(metrics["mean_throughput"])
         )
 
         logger.info("✅ Evaluación completada:")
@@ -482,7 +484,9 @@ class DQNEvaluator:
 
     def _moving_average(self, data: list[float], window: int) -> list[float]:
         """Calcular promedio móvil."""
-        return [np.mean(data[i : i + window]) for i in range(len(data) - window + 1)]
+        return [
+            float(np.mean(data[i : i + window])) for i in range(len(data) - window + 1)
+        ]
 
     def save_metrics(self) -> str:
         """
